--- Classification/classify_MAD.sh
@@ -1,9 +1,0 @@
-
-source activate mne
-cd ~/git/autism
-
-nd='--compute-nulldist'
-# irf='--short-irf'
-irf=''
-
-python classify_MAD.py --dtype $dtype $alt_tag $nd --outer-cv $outer_cv --measure $mad_type $overwrite $irf
